+++ TVShows/TVShows/Login/LoginViewController.swift
@@ -21,27 +21,10 @@ class LoginViewController: UIViewController {
     @IBOutlet weak var checkBox: UIButton!
     @IBOutlet weak var loginButton: UIButton!
     
-    var checked = UIImage(named: "ic-checkbox-filled")
-    var unchecked = UIImage(named: "ic-checkbox-empty")
-    
-    var isCheckBoxClicked:Bool!
+    private var isCheckBoxClicked:Bool!
     
     private var user: User? = nil
     private var loginToken: LoginData? = nil
-    
-    struct User: Codable {
-        let email: String
-        let type: String
-        let id: String
-        enum CodingKeys: String, CodingKey {
-            case email
-            case type
-            case id = "_id"
-        }
-    }
-    struct LoginData: Codable {
-        let token: String
-    }
     
     @IBOutlet var LabelOutlet: UILabel!
     @IBOutlet var ActivityIndicatorView: UIActivityIndicatorView!
@@ -51,7 +34,7 @@ class LoginViewController: UIViewController {
         // Do any additional setup after loading the view.
         usernameTextField.setBottomBorder()
         passwordTextField.setBottomBorder()
-        checkBox.setImage(unchecked, for: UIControlState.normal)
+        checkBox.setImage(UIImage(named: "ic-checkbox-empty"), for: UIControlState.normal)
         loginButton.roundedButton()
     }
 
@@ -74,17 +57,17 @@ class LoginViewController: UIViewController {
         
         if isCheckBoxClicked == true{
             isCheckBoxClicked = false
-            checkBox.setImage(unchecked, for: UIControlState.normal)
+            checkBox.setImage(UIImage(named: "ic-checkbox-empty"), for: UIControlState.normal)
         }
         else{
             isCheckBoxClicked = true
-            checkBox.setImage(checked, for: UIControlState.normal)
+            checkBox.setImage(UIImage(named: "ic-checkbox-filled"), for: UIControlState.normal)
         }
     }
     
     @IBAction func LoginButtonAction(_ sender: Any) {
         if let email = usernameTextField.text, !email.isEmpty, let password = passwordTextField.text, !password.isEmpty{
-            LoginAPICall(email: email, password: password)
+            loginAPICall(email: email, password: password)
         }
         
         //let storyboard = UIStoryboard(name: "Login", bundle: nil)
@@ -118,7 +101,7 @@ class LoginViewController: UIViewController {
                         switch response.result {
                             case .success(let user):
                                     self.user = user
-                                    self.LoginAPICall(email: email, password: password)
+                                    self.loginAPICall(email: email, password: password)
                             case .failure(let error):
                                     print("API failure: \(error)")
                         }
@@ -131,7 +114,7 @@ class LoginViewController: UIViewController {
         //navigationController?.pushViewController(homeViewController, animated: true)
     }
     
-    func LoginAPICall(email: String, password: String){
+    func loginAPICall(email: String, password: String){
         SVProgressHUD.show()
         
         let parameters: [String: String] = [
@@ -165,22 +148,7 @@ class LoginViewController: UIViewController {
     
 }
 
-extension UIButton{
-    
-    func roundedButton(){
-        self.layer.cornerRadius = 10
-        self.clipsToBounds = true
-    }
-}
 
-extension UITextField{
-    
-    func setBottomBorder(){
-        self.layer.shadowColor = UIColor.lightGray.cgColor
-        self.layer.shadowOffset = CGSize(width: 0.0, height: 1.0)
-        self.layer.shadowOpacity = 1.0
-        self.layer.shadowRadius = 0.0
-    }
-}
+
